// src/sanity/types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type GalleryPhoto = {
  _id: string;
  _type: "galleryPhoto";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  image?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  alt?: string;
  order?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type CATEGORIES_QUERY_RESULT = Array<{
  _id: string;
  title: string | null;
  order: number | null;
}>;

// Source: src/sanity/queries.ts
// Variable: PROJECTS_QUERY
// Query: *[_type == "project"] | order(order asc){    _id,    title,    thumbnail,    externalUrl,    role,    order,    "category": category->{ _id, title }  }
export type PROJECTS_QUERY_RESULT = Array<{
  _id: string;
  title: string | null;
  thumbnail: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
  externalUrl: string | null;
  role: string | null;
  order: number | null;
  category: {
    _id: string;
    title: string | null;
  } | null;
}>;

// Source: src/sanity/queries.ts
// Variable: GALLERY_PHOTOS_QUERY
// Query: *[_type == "galleryPhoto"] | order(order asc){    _id,    image,    alt,    order  }
export type GALLERY_PHOTOS_QUERY_RESULT = Array<{
  _id: string;
  image: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
  alt: string | null;
  order: number | null;
}>;